Refactored several scripts
Added additional logging information
