+++ Unity/Assets/RealityFlow/Scripting/RoslynCodeRunner.cs
@@ -7,6 +7,7 @@ using System.Linq;
 using System.Reflection;
 using UnityEngine;
 using UnityEngine.Events;
+using System.IO;
 
 public class RoslynCodeRunner : Singleton<RoslynCodeRunner>
 {
@@ -31,6 +32,13 @@ public class RoslynCodeRunner : Singleton<RoslynCodeRunner>
     private string resultInfo;
 
     private readonly List<Diagnostic> diagnostics = new List<Diagnostic>();
+
+    private string logFilePath;
+
+    private void Start()
+    {
+        logFilePath = Path.Combine(Application.persistentDataPath, "ChatGPTLogs.txt");
+    }
 
     public IEnumerator RunCodeCoroutine(string updatedCode = null)
     {
@@ -73,7 +81,8 @@ public class RoslynCodeRunner : Singleton<RoslynCodeRunner>
         // Check if the compilation resulted in a valid assembly
         if (asm == null)
         {
-            Logger.Instance.LogError("There was an error completing this action try submiting your request again...");
+            LogScriptExecutionResult("Failed to compile code.");
+            Logger.Instance.LogError("There was an error completing this action try submitting your request again...");
             Debug.LogError("Failed to compile code.");
             yield break;
         }
@@ -82,6 +91,7 @@ public class RoslynCodeRunner : Singleton<RoslynCodeRunner>
         Type type = asm.GetTypes().Where(t => t.GetMethod("Execute", BindingFlags.Static | BindingFlags.Public) != null).FirstOrDefault();
         if (type == null)
         {
+            LogScriptExecutionResult("No static Execute method found.");
             Debug.LogError("No static Execute method found.");
             yield break;
         }
@@ -90,6 +100,7 @@ public class RoslynCodeRunner : Singleton<RoslynCodeRunner>
         MethodInfo method = type.GetMethod("Execute", BindingFlags.Static | BindingFlags.Public);
         if (method == null)
         {
+            LogScriptExecutionResult("Execute method not found.");
             Debug.LogError("Execute method not found.");
             yield break;
         }
@@ -101,10 +112,12 @@ public class RoslynCodeRunner : Singleton<RoslynCodeRunner>
             try
             {
                 method.Invoke(null, null);
+                LogScriptExecutionResult("Script executed successfully.");
             }
             catch (Exception e)
             {
                 Debug.LogError("Error during script execution: " + e.Message);
+                LogScriptExecutionResult($"Error during script execution: {e.Message}");
             }
             methodInvocationCompleted = true;
         });
@@ -116,7 +129,9 @@ public class RoslynCodeRunner : Singleton<RoslynCodeRunner>
         }
     }
 
-
-
-
+    private void LogScriptExecutionResult(string message)
+    {
+        string logMessage = $"[{DateTime.Now}] Script Execution Result: {message}\n";
+        File.AppendAllText(logFilePath, logMessage);
+    }
 }
