+++ Unity/Assets/RealityFlow/Third Party/UnityChatGPT/ChatGPTReminders.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 000f2af9a7b922e468293dfab8bf3f90
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
